requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.0
filelock==3.9.0
Flask==2.2.2
huggingface-hub==0.12.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keybert==0.7.0
llvmlite==0.39.1
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pyarrow==11.0.0
Pygments==2.14.0
pynndescent==0.5.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@d60ce38b6a2d8aa55f29502f5517d47c1d9e4a5c
soupsieve==2.3.2.post1
tenacity==8.2.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.14
Werkzeug==2.2.2
zope.interface==5.5.2
